Remove the ZIR load instruction to preserve the name
Previously the load instruction was used as a wrapper (or actually
the owner) of immediate values and identifiers. Instead of that, we
can let each -variable or cast- instruction to own the operands to know
when they should free the memory. In the same way, the operands that contain
references to other operands can do the same.