Implement a RP2A03 program
The new module is in charge of generating machine code for the RP2A03
processor and can also disassemble programs. There is an internal
representation of a NES program that is generated from a ZIR program
that is used to create the RP2A03 program and it can also be used to
transpile to different assemblers like ca65 or ASM6.